fixed target assignment api
